refactor: Split reducers based on their functionality fix: Improve how components access the state in given reducer feat: Create 'Add to cart' and 'Add to wishlist' buttons feat: Facilitate control elements to customize product order feat: Animate cart when it recieves a product 😎 feat: Improve UI for product customization controls feat: Create an overlay component to darken screen for modals feat: Create cart button in header fix: Change tx_id column to accept decimal instead of big integer refactor: Parse form data into json in OrderController feat: Integrate paytm checkout 🔥

+++ database/migrations/2019_05_10_180918_create_orders_table.php
@@ -19,7 +19,7 @@ class CreateOrdersTable extends Migration
             $table->decimal('total',12,2);
             $table->decimal('saved',12,2);
             $table->boolean('tx_status')->nullable();
-            $table->integer('tx_id')->nullable();
+            $table->decimal('tx_id',36,0)->nullable();
             $table->mediumText('err_mssg');
             $table->dateTime('payment_date');
